creacion

+++ src/main/java/com/tecsup/petclinic/services/VisitService.java
@@ -7,59 +7,11 @@ import java.sql.Date;
 import java.util.List;
 
 public interface VisitService {
-    /**
-     * Crea una nueva visita
-     *
-     * @param visit
-     * @return
-     */
     Visit create(Visit visit);
-
-    /**
-     * Actualiza una visita existente
-     *
-     * @param visit
-     * @return
-     */
     Visit update(Visit visit);
-
-    /**
-     * Elimina una visita por su ID
-     *
-     * @param id
-     * @throws VisitNotFoundException
-     */
-    void delete(Long id) throws VisitNotFoundException;
-
-    /**
-     * Busca una visita por su ID
-     *
-     * @param id
-     * @return
-     * @throws VisitNotFoundException
-     */
-    Visit findById(Long id) throws VisitNotFoundException;
-
-    /**
-     * Busca visitas por el ID de la mascota
-     *
-     * @param petId
-     * @return
-     */
-    List<Visit> findByPetId(Long petId);
-
-    /**
-     * Busca visitas por la fecha de la visita
-     *
-     * @param visitDate
-     * @return
-     */
+    void delete(Integer id) throws VisitNotFoundException;
+    Visit findById(Integer id) throws VisitNotFoundException;
+    List<Visit> findByPetId(Integer petId);
     List<Visit> findByVisitDate(Date visitDate);
-
-    /**
-     * Lista todas las visitas
-     *
-     * @return
-     */
     List<Visit> findAll();
 }
